requirements.txt
alembic==1.7.5; python_version >= "3.6"
bidict==0.21.4; python_version >= "3.6"
cachelib==0.6.0; python_version >= "3.6"
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.6.2" and python_full_version < "4.0.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0")
commonmark==0.9.1; python_full_version >= "3.6.2" and python_full_version < "4.0.0"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
dnspython==2.2.0; python_version >= "3.6" and python_version < "4.0"
easyai==2.0.12
eventlet==0.33.0
flask-assets==2.0
flask-migrate==3.1.0; python_version >= "3.6"
flask-session==0.4.0
flask-socketio==5.1.1; python_version >= "3.6"
flask-sqlalchemy==2.5.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flask-wtf==1.0.0; python_version >= "3.6"
flask==2.0.2; python_version >= "3.6"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
gunicorn==20.1.0; python_version >= "3.5"
importlib-metadata==3.10.1; python_version < "3.9" and python_version >= "3.6"
importlib-resources==5.4.0; python_version < "3.9" and python_version >= "3.6"
itsdangerous==2.0.1; python_version >= "3.6"
jinja2==3.0.3; python_version >= "3.6"
mako==1.1.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
mctspy==0.1.1; python_full_version >= "3.5.7"
numpy==1.22.1; python_version >= "3.8" and python_full_version >= "3.5.7"
packaging==21.3; python_version >= "3.6"
pygments==2.11.2; python_version >= "3.5"
pyparsing==3.0.7; python_version >= "3.6"
python-dotenv==0.19.2; python_version >= "3.5"
python-engineio==4.3.1; python_version >= "3.6"
python-socketio==5.5.1; python_version >= "3.6"
redis==4.1.1; python_version >= "3.6"
rich==11.0.0; python_full_version >= "3.6.2" and python_full_version < "4.0.0"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlalchemy==1.4.31; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
structlog==21.5.0; python_version >= "3.6"
webassets==2.0
werkzeug==2.0.2; python_version >= "3.6"
wrapt==1.13.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
wtforms==3.0.1; python_version >= "3.7"
zipp==3.7.0; python_version < "3.9" and python_version >= "3.7"
